+++ packages/jest-runtime/src/__tests__/script_transformer.test.js
@@ -11,7 +11,8 @@
 jest
   .mock('fs', () =>
     // Node 10.5.x compatibility
-    Object.assign({}, jest.genMockFromModule('fs'), {
+    ({
+      ...jest.genMockFromModule('fs'),
       ReadStream: jest.requireActual('fs').ReadStream,
       WriteStream: jest.requireActual('fs').WriteStream,
       readFileSync: jest.fn((path, options) => {
@@ -264,9 +265,10 @@ describe('ScriptTransformer', () => {
     "throws an error if `process` doesn't return a string or an object" +
       'containing `code` key with processed string',
     () => {
-      config = Object.assign(config, {
+      config = {
+        ...config,
         transform: [['^.+\\.js$', 'passthrough-preprocessor']],
-      });
+      };
       const scriptTransformer = new ScriptTransformer(config);
 
       const incorrectReturnValues = [
@@ -299,9 +301,10 @@ describe('ScriptTransformer', () => {
   );
 
   it("throws an error if `process` doesn't defined", () => {
-    Object.assign(config, {
+    config = {
+      ...config,
       transform: [['^.+\\.js$', 'skipped-required-props-preprocessor']],
-    });
+    };
     const scriptTransformer = new ScriptTransformer(config);
     expect(() =>
       scriptTransformer.transformSource('sample.js', '', false),
@@ -309,11 +312,12 @@ describe('ScriptTransformer', () => {
   });
 
   it('throws an error if createTransformer returns object without `process` method', () => {
-    Object.assign(config, {
+    config = {
+      ...config,
       transform: [
         ['^.+\\.js$', 'skipped-required-create-transformer-props-preprocessor'],
       ],
-    });
+    };
     const scriptTransformer = new ScriptTransformer(config);
     expect(() =>
       scriptTransformer.transformSource('sample.js', '', false),
@@ -321,9 +325,10 @@ describe('ScriptTransformer', () => {
   });
 
   it("shouldn't throw error without process method. But with corrent createTransformer method", () => {
-    Object.assign(config, {
+    config = {
+      ...config,
       transform: [['^.+\\.js$', 'skipped-process-method-preprocessor']],
-    });
+    };
     const scriptTransformer = new ScriptTransformer(config);
     expect(() =>
       scriptTransformer.transformSource('sample.js', '', false),
@@ -331,9 +336,7 @@ describe('ScriptTransformer', () => {
   });
 
   it('uses the supplied preprocessor', () => {
-    config = Object.assign(config, {
-      transform: [['^.+\\.js$', 'test_preprocessor']],
-    });
+    config = {...config, transform: [['^.+\\.js$', 'test_preprocessor']]};
     const scriptTransformer = new ScriptTransformer(config);
     scriptTransformer.transform('/fruits/banana.js', {});
 
@@ -347,12 +350,13 @@ describe('ScriptTransformer', () => {
   });
 
   it('uses multiple preprocessors', () => {
-    config = Object.assign(config, {
+    config = {
+      ...config,
       transform: [
         ['^.+\\.js$', 'test_preprocessor'],
         ['^.+\\.css$', 'css-preprocessor'],
       ],
-    });
+    };
     const scriptTransformer = new ScriptTransformer(config);
 
     scriptTransformer.transform('/fruits/banana.js', {});
@@ -369,9 +373,10 @@ describe('ScriptTransformer', () => {
   });
 
   it('writes source map if preprocessor supplies it', () => {
-    config = Object.assign(config, {
+    config = {
+      ...config,
       transform: [['^.+\\.js$', 'preprocessor-with-sourcemaps']],
-    });
+    };
     const scriptTransformer = new ScriptTransformer(config);
 
     const map = {
@@ -395,9 +400,10 @@ describe('ScriptTransformer', () => {
   });
 
   it('writes source map if preprocessor inlines it', () => {
-    config = Object.assign(config, {
+    config = {
+      ...config,
       transform: [['^.+\\.js$', 'preprocessor-with-sourcemaps']],
-    });
+    };
     const scriptTransformer = new ScriptTransformer(config);
 
     const sourceMap = JSON.stringify({
@@ -424,9 +430,10 @@ describe('ScriptTransformer', () => {
   });
 
   it('writes source maps if given by the transformer', () => {
-    config = Object.assign(config, {
+    config = {
+      ...config,
       transform: [['^.+\\.js$', 'preprocessor-with-sourcemaps']],
-    });
+    };
     const scriptTransformer = new ScriptTransformer(config);
 
     const map = {
@@ -453,9 +460,10 @@ describe('ScriptTransformer', () => {
   });
 
   it('does not write source map if not given by the transformer', () => {
-    config = Object.assign(config, {
+    config = {
+      ...config,
       transform: [['^.+\\.js$', 'preprocessor-with-sourcemaps']],
-    });
+    };
     const scriptTransformer = new ScriptTransformer(config);
 
     require('preprocessor-with-sourcemaps').process.mockReturnValue({
@@ -471,9 +479,7 @@ describe('ScriptTransformer', () => {
   });
 
   it('passes expected transform options to getCacheKey', () => {
-    config = Object.assign(config, {
-      transform: [['^.+\\.js$', 'test_preprocessor']],
-    });
+    config = {...config, transform: [['^.+\\.js$', 'test_preprocessor']]};
     const scriptTransformer = new ScriptTransformer(config);
 
     scriptTransformer.transform('/fruits/banana.js', {
@@ -485,10 +491,8 @@ describe('ScriptTransformer', () => {
   });
 
   it('reads values from the cache', () => {
-    const transformConfig = Object.assign(config, {
-      transform: [['^.+\\.js$', 'test_preprocessor']],
-    });
-    let scriptTransformer = new ScriptTransformer(transformConfig);
+    config = {...config, transform: [['^.+\\.js$', 'test_preprocessor']]};
+    let scriptTransformer = new ScriptTransformer(config);
     scriptTransformer.transform('/fruits/banana.js', {});
 
     const cachePath = getCachePath(mockFs, config);
@@ -502,7 +506,7 @@ describe('ScriptTransformer', () => {
 
     // Restore the cached fs
     mockFs = mockFsCopy;
-    scriptTransformer = new ScriptTransformer(transformConfig);
+    scriptTransformer = new ScriptTransformer(config);
     scriptTransformer.transform('/fruits/banana.js', {});
 
     expect(fs.readFileSync.mock.calls.length).toBe(2);
@@ -514,8 +518,8 @@ describe('ScriptTransformer', () => {
     jest.resetModuleRegistry();
     reset();
     mockFs = mockFsCopy;
-    transformConfig.cache = false;
-    scriptTransformer = new ScriptTransformer(transformConfig);
+    config.cache = false;
+    scriptTransformer = new ScriptTransformer(config);
     scriptTransformer.transform('/fruits/banana.js', {});
 
     expect(fs.readFileSync.mock.calls.length).toBe(1);
@@ -525,19 +529,17 @@ describe('ScriptTransformer', () => {
   });
 
   it('does not reuse the in-memory cache between different projects', () => {
-    const scriptTransformer = new ScriptTransformer(
-      Object.assign({}, config, {
-        transform: [['^.+\\.js$', 'test_preprocessor']],
-      }),
-    );
+    const scriptTransformer = new ScriptTransformer({
+      ...config,
+      transform: [['^.+\\.js$', 'test_preprocessor']],
+    });
 
     scriptTransformer.transform('/fruits/banana.js', {});
 
-    const anotherScriptTransformer = new ScriptTransformer(
-      Object.assign({}, config, {
-        transform: [['^.+\\.js$', 'css-preprocessor']],
-      }),
-    );
+    const anotherScriptTransformer = new ScriptTransformer({
+      ...config,
+      transform: [['^.+\\.js$', 'css-preprocessor']],
+    });
 
     anotherScriptTransformer.transform('/fruits/banana.js', {});